[WIP] Integrate CMake and implement tokenizer
CMake finally works after hours of struggles! Woohoo!

Implementation of the core of the plagiarism detection engine has
begun. Because various algorithms, such as MOSS's winnowing, require
tokenization to process and compare input files, writing a tokenizer
has been chosen as a starting point. This tokenizer receives a
filename to tokenize and the file's source language. Then, it
produces three vectors of tokens: one of code tokens (strings),
one of integer mappings from the code tokens, and one of masked
integer mappings (e.g., all identifiers are assigned the same number
rather than distinct numbers).

Unfortunately... The tokenizer is still a work in progress. When
testing the tokenizer to see the tokens it would produce, the
process seems to get stuck. This appears to occur when the tokenizer
is calling the language-specific tokenizers from dspinellis's
tokenizer repo. A good next step would be tracing the function calls
to see where exactly things get stuck / start looping.